requirements.txt
sentence-transformers
faiss-cpu
numpy
torch
requests